Still WIP, but show thumbnails per instance in series view
Have a ui widget factory plugin that is configured to show
the images for a series.

index.html is now parameterized by the series ID to show and it
creates the series display based on what is passed in.

TODO: make it so that clicking on a series in the root view redirects
to a url with the series uid as a parameter.