WIP: Add an a new control point overlay experiment
* the database can now contain documents containing
sets of per-instance control points for curves
* when an instance is displayed the control points
for that instance are fetched
* the slice view shows the control points in the svn overlay

Currently the points are hard-coded from an instance known to have
points.

The svg points are editable, but not as curves.

The control points are pulled from the database before the
drawing is complete, so there is more lag than needed.
Would be better to grab all control points for a series
into the client and draw then as the instance changes.