Implement control point access as instance of 'report-like' class
Now when a series is selected the control points are all downloaded
and used to display the overlays.

The control points document is now treated as an instance of an
as-yet undefined class.  In future Chronicle, the control points
should be stored in an instance of structured report of the
correct template.  Probably it's enough to know that derived data
is an SR instance, but it might be good to consider putting
the instanced templates also in the view.